requirements.txt
flet
supabase
cloudinary
bcrypt
gradio_client
python-dotenv
google-generativeai
requests
Pillow